almost finished

+++ database/migrations/2024_04_21_151938_create_user_carts_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->Time('jobstarttime')->nullable();
             $table->Time('jobendtime')->nullable();
             $table->String('completed');
-            $table->Integer('cost')->nullable();
+            $table->Float('cost')->nullable();
             $table->timestamps();
         });
     }
